pyproject.toml
[tool.poetry]
name = "llm-local"
version = "0.1.0"
description = ""
authors = ["Maksym Stepanenko"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
transformers = "^4.48.2"
accelerate = "^1.3.0"
huggingface-hub = "^0.28.1"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
mypy = "^1.14.1"
ruff = "^0.9.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
files = "llm_local"
ignore_missing_imports = true

[tool.poetry.scripts]
app = "llm_local.app:main"
